requirements.txt
torch>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pyyaml>=6.0
xarray>=2023.1.0
tqdm>=4.65.0
